@@ -8,13 +8,13 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "ramda": "^0.23.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2",
     "react-redux": "^5.0.3",
     "redux": "^3.6.0",
-    "redux-act": "^1.1.1",
+    "redux-act": "^1.5.0",
     "redux-devtools-extension": "^2.13.0",
-    "redux-persist": "^4.4.2",
+    "redux-persist": "^4.10.2",
     "redux-sounds": "^1.1.1"
   },
   "scripts": {
